+++ addon/controllers/fd-editform-constructor.js
@@ -186,12 +186,15 @@ export default Ember.Controller.extend(FdWorkPanelToggler, {
     }
 
     let dataobjectId = this.get('model.dataobject.id');
-    let view = this.get('model.editform.formViews.firstObject.view');
+    let mockView = Ember.Object.create({
+      definition: controlsToDefinition(this.get('controlsTree')),
+      class: this.get('model.dataobject')
+    });
 
     let dataForBuildTree = getDataForBuildTree(this.get('store'), dataobjectId);
-    let attributesForTree = getTreeNodeByNotUsedAttributes(this.get('store'), dataForBuildTree.classes, view, 'type');
+    let attributesForTree = getTreeNodeByNotUsedAttributes(this.get('store'), dataForBuildTree.classes, mockView, 'type');
     let associationForTree = getAssociationTreeNode(Ember.A(), dataForBuildTree.associations, 'node_', dataobjectId, 'name');
-    let aggregationForTree = getTreeNodeByNotUsedAggregation(dataForBuildTree.aggregations, view, 'name');
+    let aggregationForTree = getTreeNodeByNotUsedAggregation(dataForBuildTree.aggregations, mockView, 'name');
 
     let attributesTree = Ember.A();
     attributesTree.pushObjects([
@@ -386,7 +389,7 @@ export default Ember.Controller.extend(FdWorkPanelToggler, {
     @type Ember.NativeArray
   */
   simpleTypes: Ember.computed('_dataTypes', function() {
-    return this._buildTree(this.get('_dataTypes').flexberryTypes(), 'property');
+    return this._buildTree(this.get('_dataTypes').fDTypes(), 'property');
   }),
 
   /**
@@ -421,7 +424,7 @@ export default Ember.Controller.extend(FdWorkPanelToggler, {
     @property dataObjectProperties
     @type Object
   */
-  dataObjectProperties: Ember.computed('model.dataobject', 'model.inheritances', 'model.associations', 'model.aggregations', function() {
+  dataObjectProperties: Ember.computed('model.dataobject.attributes.@each.type', 'model.inheritances', 'model.associations', 'model.aggregations', function() {
     let dataObject = this.get('model.dataobject');
     let inheritances = this.get('model.inheritances');
     let associations = this.get('model.associations');
@@ -475,13 +478,10 @@ export default Ember.Controller.extend(FdWorkPanelToggler, {
 
       this.notifyPropertyChange('_dataForBuildTree');
 
-      let view = this.get('model.editform.formViews.firstObject.view');
-      let viewDefinition = view.get('definition');
       let propertyDefinition = FdViewAttributesProperty.create({
         name: 'newAttribute' + atrIndex,
         visible: true,
       });
-      viewDefinition.pushObject(propertyDefinition);
 
       let control = FdEditformControl.create({
         caption: `${this.get('i18n').t('forms.fd-editform-constructor.new-control-caption').toString()} #${this.incrementProperty('_newControlIndex')}`,
@@ -549,8 +549,9 @@ export default Ember.Controller.extend(FdWorkPanelToggler, {
         this._removeItem(this.get('selectedItem'));
 
         // Refresh definition for filter not used attributes in 'dataNotUsedAttributesTreeObserver'.
-        let view = this.get('model.editform.formViews.firstObject.view');
-        view.set('definition', controlsToDefinition(this.get('controlsTree')));
+        // let view = this.get('model.editform.formViews.firstObject.view');
+        // view.set('definition', controlsToDefinition(this.get('controlsTree')));
+
         this.set('selectedItem', undefined);
       } else {
         this.set('_showModalDialog', true);
@@ -603,9 +604,16 @@ export default Ember.Controller.extend(FdWorkPanelToggler, {
           }
         }
       } else if (!this.get('_moveItem')) {
-        let newSelectedItem = selectedItem === item ? undefined : item;
-        this.set('selectedItem', newSelectedItem);
+        let configPanelSidebar = Ember.$('.ui.sidebar.config-panel');
+        let sidebarOpened = configPanelSidebar.hasClass('visible');
 
+        if ((item || sidebarOpened) && selectedItem !== item) {
+          this.send('toggleConfigPanel', 'control-properties', item);
+        }
+
+        this.set('selectedItem', item);
+
+        let newSelectedItem = selectedItem === item ? undefined : item;
         if (!Ember.isNone(newSelectedItem) && newSelectedItem.get('propertyDefinition.name') === '') {
           this.set('_showNotUsedAttributesTree', true);
         } else {
@@ -1122,7 +1130,7 @@ export default Ember.Controller.extend(FdWorkPanelToggler, {
       let { aggregations } = dataObjectProperties;
       let relation = aggregations.findBy('endRole', propertyName) || aggregations.findBy('endClass.name', propertyName);
       type = 'detail';
-      view = this.get('model.views').filterBy('class.id', relation.get('endClass.id')).findBy('name', propertyDefinition.get('detailViewName'));
+      view = Ember.A(this.get('model.views').filterBy('class.id', relation.get('endClass.id'))).findBy('name', propertyDefinition.get('detailViewName'));
       types = this._getTypesForView(view);
     } else if (propertyDefinition instanceof FdViewAttributesMaster) {
       type = 'master';
@@ -1142,22 +1150,29 @@ export default Ember.Controller.extend(FdWorkPanelToggler, {
       while (path.length > 0) {
         let role = path.shift();
         let relation = properties.associations.findBy('startRole', role) || properties.associations.findBy('startClass.name', role);
-        properties = this._getClassProperties(relation.get('startClass'), inheritances, associations, aggregations);
+        if (relation) {
+          properties = this._getClassProperties(relation.get('startClass'), inheritances, associations, aggregations);
+        } else {
+          console.error('Not found association with name:' + role);
+        }
       }
 
       let attribute = properties.attributes.findBy('name', propertyName);
-      let typeInMap = this.get('model.stage.typeMapCS').findBy('name', attribute.get('type'));
-
-      if (typeInMap) {
-        type = typeInMap.value || typeInMap.name;
-      } else {
-        let clazz = this.get('model.classes').findBy('name', attribute.get('type'));
-        if (clazz && clazz.get('stereotype') === '«enumeration»') {
-          type = 'enumeration';
-          items = clazz.get('attributes').mapBy('name');
+      if (attribute) {
+        let typeInMap = this.get('model.stage.typeMapCS').findBy('name', attribute.get('type'));
+        if (typeInMap) {
+          type = typeInMap.value || typeInMap.name;
         } else {
-          type = 'default';
+          let clazz = this.get('model.classes').findBy('name', attribute.get('type'));
+          if (clazz && clazz.get('stereotype') === '«enumeration»') {
+            type = 'enumeration';
+            items = clazz.get('attributes').mapBy('name');
+          } else {
+            type = 'default';
+          }
         }
+      } else {
+        type = 'default';
       }
     } else {
       throw new Error('Invalid property definition.');